requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.7
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
flower==2.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
idna==3.10
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pylint==3.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==4.6.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
time-machine==2.16.0
tomlkit==0.13.2
tornado==6.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
